requirements.txt
streamlit
pillow
numpy
pandas
matplotlib
plotly
scikit-learn
